feat: port를 사용하여 순환 참조 구조 개선
