--- BE/src/main/java/roomescape/reservation/application/ReservationReferenceChecker.java
@@ -1,6 +1,0 @@
-package roomescape.reservation.application;
-
-public interface ReservationReferenceChecker {
-    void validateReservationTimeNotReferenced(Long timeId);
-    void validateThemeNotReferenced(Long themeId);
-}
