+++ BE/src/main/java/roomescape/reservation/application/ThemeReferenceChecker.java
@@ -1,25 +1,18 @@
 package roomescape.reservation.application;
 
 import org.springframework.stereotype.Component;
-import roomescape.global.exception.ReservationTimeErrorCode;
 import roomescape.global.exception.ThemeErrorCode;
 import roomescape.global.exception.customException.BusinessException;
 import roomescape.reservation.domain.ReservationRepository;
+import roomescape.theme.application.ReferenceChecker;
 
 @Component
-public class DefaultReferenceChecker implements ReservationReferenceChecker {
+public class ThemeReferenceChecker implements ReferenceChecker {
 
     private final ReservationRepository reservationRepository;
 
-    public DefaultReferenceChecker(ReservationRepository reservationRepository) {
+    public ThemeReferenceChecker(ReservationRepository reservationRepository) {
         this.reservationRepository = reservationRepository;
-    }
-
-    @Override
-    public void validateReservationTimeNotReferenced(Long timeId) {
-        if (reservationRepository.existsByReservationTimeId(timeId)) {
-            throw new BusinessException(ReservationTimeErrorCode.RESERVATION_TIME_IN_USE);
-        }
     }
 
     @Override
