+++ BE/src/main/java/roomescape/reservation/domain/Reservation.java
@@ -1,9 +1,6 @@
 package roomescape.reservation.domain;
 
 import java.time.LocalDate;
-import java.time.LocalDateTime;
-import roomescape.global.exception.ReservationErrorCode;
-import roomescape.global.exception.customException.BusinessException;
 import roomescape.reservationTime.domain.ReservationTime;
 import roomescape.theme.domain.Theme;
