+++ BE/src/main/java/roomescape/reservationTime/application/ReferenceChecker.java
@@ -1,0 +1,5 @@
+package roomescape.reservationTime.application;
+
+public interface ReferenceChecker {
+    void validateReservationTimeNotReferenced(Long timeId);
+}
