+++ BE/src/main/java/roomescape/theme/application/ReferenceChecker.java
@@ -1,0 +1,5 @@
+package roomescape.theme.application;
+
+public interface ReferenceChecker {
+    void validateThemeNotReferenced(Long themeId);
+}
