+++ BE/src/test/java/roomescape/reservationTime/application/ReservationTimeServiceTest.java
@@ -12,8 +12,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.global.exception.customException.BusinessException;
 import roomescape.global.exception.customException.EntityNotFoundException;
-import roomescape.reservation.application.DefaultReferenceChecker;
-import roomescape.reservation.application.ReservationReferenceChecker;
+import roomescape.reservation.application.ReservationTimeReferenceChecker;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.fake.FakeReservationRepository;
@@ -28,14 +27,14 @@ class ReservationTimeServiceTest {
     private ReservationTimeRepository reservationTimeRepository;
     private ReservationRepository reservationRepository;
     private ReservationTimeService reservationTimeService;
-    private ReservationReferenceChecker referenceChecker;
+    private ReferenceChecker referenceChecker;
 
 
     @BeforeEach
     void setUp() {
         reservationTimeRepository = new FakeReservationTimeRepository();
         reservationRepository = new FakeReservationRepository();
-        referenceChecker = new DefaultReferenceChecker(reservationRepository);
+        referenceChecker = new ReservationTimeReferenceChecker(reservationRepository);
         reservationTimeService = new ReservationTimeService(
                 reservationTimeRepository,
                 reservationRepository,
