+++ BE/src/test/java/roomescape/theme/application/ThemeServiceTest.java
@@ -12,8 +12,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import roomescape.global.exception.customException.BusinessException;
 import roomescape.global.exception.customException.EntityNotFoundException;
-import roomescape.reservation.application.DefaultReferenceChecker;
-import roomescape.reservation.application.ReservationReferenceChecker;
+import roomescape.reservation.application.ThemeReferenceChecker;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.domain.ReservationRepository;
 import roomescape.reservation.fake.FakeReservationRepository;
@@ -36,7 +35,7 @@ class ThemeServiceTest {
     private PopularThemeRepository popularThemeRepository;
     private ThemeService themeService;
     private ReservationRepository reservationRepository;
-    private ReservationReferenceChecker referenceChecker;
+    private ReferenceChecker referenceChecker;
 
     @BeforeEach
     void setUp() {
@@ -47,7 +46,7 @@ class ThemeServiceTest {
                 themeRepository.findAll()
         );
         reservationRepository = new FakeReservationRepository();
-        referenceChecker = new DefaultReferenceChecker(reservationRepository);
+        referenceChecker = new ThemeReferenceChecker(reservationRepository);
         themeService = new ThemeService(referenceChecker, themeRepository, popularThemeRepository);
     }
